requirements.txt
requests>=2.31.0
beautifulsoup4>=4.12.0
selenium>=4.15.0
PyPDF2>=3.0.0
pdfplumber>=0.10.0
PyMuPDF>=1.23.0
pandas>=2.0.0
numpy>=1.24.0
openai>=1.0.0
python-dotenv>=1.0.0
fuzzywuzzy[speedup]>=0.18.0
nltk>=3.8.0
aiohttp>=3.9.0
tqdm>=4.65.0
jsonschema>=4.17.0